Change tracker-humanity into HumanityCheck.
