+++ src/utils/humanity-check.js
@@ -1,0 +1,53 @@
+const store = require('../utils/store');
+const botList = require('./botlist');
+
+const storageKey = '_constructorio_is_human';
+const humanEvents = [
+  'scroll',
+  'resize',
+  'touchmove',
+  'mouseover',
+  'mousemove',
+  'keydown',
+  'keypress',
+  'keyup',
+  'focus',
+];
+
+class HumanityCheck {
+  constructor() {
+    this.isHumanBoolean = !!store.session.get(storageKey) || false;
+
+    // Humanity proved, remove handlers to prove humanity
+    const remove = () => {
+      this.isHumanBoolean = true;
+
+      store.session.set(storageKey, true);
+      humanEvents.forEach((eventType) => {
+        window.removeEventListener(eventType, remove, true);
+      });
+    };
+
+    // Add handlers to prove humanity
+    if (!this.isHumanBoolean) {
+      humanEvents.forEach((eventType) => {
+        window.addEventListener(eventType, remove, true);
+      });
+    }
+  }
+
+  // Return boolean indicating if is human
+  isHuman() {
+    return this.isHumanBoolean || !!store.session.get(storageKey);
+  }
+
+  // Return boolean indicating if useragent matches botlist
+  isBot() {
+    const { userAgent, webdriver } = window && window.navigator;
+    const botRegex = new RegExp(`(${botList.join('|')})`);
+
+    return Boolean(userAgent.match(botRegex)) || Boolean(webdriver);
+  }
+};
+
+module.exports = HumanityCheck;
